requirements.txt
boto3==1.34.162
botocore==1.34.162
yfinance==0.2.37
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
requests==2.32.3
dataclasses-json==0.6.7
pydantic==2.9.2
pandas==2.2.3
numpy==1.24.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1